Renamed EditorFrame -> index
